refactor: extract employee BWR panel component
